#10071: Move test files to TTNN
