+++ tests/ttnn/unit_tests/operations/backward/test_backward_log.py
@@ -5,7 +5,7 @@
 import torch
 import pytest
 import ttnn
-from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import (
+from tests.ttnn.unit_tests.operations.backward.utility_funcs import (
     data_gen_with_val,
     compare_pcc,
     data_gen_with_range,
