+++ tests/ttnn/unit_tests/operations/backward/test_backward_round.py
@@ -5,7 +5,7 @@
 import torch
 import pytest
 import ttnn
-from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import data_gen_with_range, compare_pcc
+from tests.ttnn.unit_tests.operations.backward.utility_funcs import data_gen_with_range, compare_pcc
 
 
 @pytest.mark.parametrize(
@@ -16,19 +16,17 @@ from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs i
         (torch.Size([1, 3, 320, 384])),
     ),
 )
-def test_bw_abs(input_shapes, device):
-    in_data, input_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
-    grad_data, grad_tensor = data_gen_with_range(input_shapes, -100, 100, device)
+def test_bw_round(input_shapes, device):
+    grad_data, grad_tensor = data_gen_with_range(input_shapes, -100, 199, device)
+    in_data, input_tensor = data_gen_with_range(input_shapes, -200, 201, device, required_grad=True)
+    pyt_y = torch.round(in_data)
 
-    pyt_y = torch.abs(in_data)
-
-    tt_output_tensor_on_device = ttnn.abs_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.round_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
 
     pyt_y.backward(gradient=grad_data)
 
     golden_tensor = [in_data.grad]
-
     comp_pass = compare_pcc(tt_output_tensor_on_device, golden_tensor)
     assert comp_pass
